+++ src/error.rs
@@ -16,4 +16,8 @@ pub enum Error {
 
     /// The input doesn't yield enough input elements.
     TooFewInputElements,
+
+    /// Failed to decrypt the message from the cipher with the provided secret
+    /// and nonce.
+    DecryptionFailed,
 }